# docs/api-specification.yml
openapi: 3.0.3
info:
  title: テストケース管理システム API
  description: |
    Next.js (App Router) + PostgreSQL + NextAuth.jsを使用したテストケース管理システムのREST API仕様書

    ## 認証
    すべてのAPIエンドポイント（認証関連を除く）はJWT認証が必要です。
    NextAuth.jsのセッショントークンを使用してリクエストを行います。

    ## 権限管理
    ### 静的ロール (user_role)
    - 0: 管理者 - すべての操作が可能
    - 1: テスト管理者 - テストグループの作成が可能
    - 2: 一般 - 割り当てられたテストグループのみアクセス可能

    ### 動的ロール (test_role)
    - 1: 設計者 - テストケースの作成・編集が可能
    - 2: 実施者 - テスト結果の登録が可能
    - 3: 閲覧者 - 閲覧のみ

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: ユーザー一覧取得
      description: |
        全ユーザーの一覧を取得（管理者のみ）
        **実装状態**: ✅ 完了
      operationId: getUsers
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          description: メールアドレスで検索（部分一致）
          schema:
            type: string
        - name: department
          in: query
          description: 部署で検索（部分一致）
          schema:
            type: string
        - name: tagId
          in: query
          description: タグIDで検索
          schema:
            type: integer
      responses:
        '200':
          description: ユーザー一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithTags'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Users
      summary: ユーザー作成
      description: |
        新しいユーザーを作成（管理者のみ）
        - タグが存在しない場合は自動作成
        **実装状態**: ✅ 完了
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth.js JWT トークン
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        user_role:
          type: integer
        department:
          type: string
        company:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserWithTags:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            tags:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
    UserInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        user_role:
          type: integer
          default: 2
        department:
          type: string
        company:
          type: string
        tags:
          type: array
          description: タグ名の配列（存在しない場合は自動作成）
          items:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
  responses:
    BadRequest:
      description: リクエストが不正です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 権限がありません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []